+++ packages/rax-plugin-app/src/config/user/default.config.js
@@ -1,6 +1,18 @@
+const address = require('address');
+
 module.exports = {
   inlineStyle: true,
   outputDir: 'build',
   publicPath: '/',
   devPublicPath: '/',
+  devServer: {
+    compress: true,
+    clientLogLevel: 'error',
+    historyApiFallback: true,
+    hot: true,
+    quiet: true,
+    overlay: false,
+    host: address.ip(),
+    port: 9999,
+  }
 };
