+++ packages/rax-plugin-app/src/config/web/setBuild.js
@@ -1,29 +1,7 @@
 'use strict';
 
-const path = require('path');
-const UglifyJSPlugin = require('uglifyjs-webpack-plugin');
-const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
-
-const UNIVERSAL_APP_SHELL_LOADER = require.resolve('universal-app-shell-loader');
+const setUserConfig = require('../user/setConfig');
 
 module.exports = (config, context) => {
-  const { rootDir } = context;
-  const appEntry = path.resolve(rootDir, 'src/app.js');
-
-  config.mode('production');
-  config.devtool('source-map');
-  config.entry('index')
-    .add(`${UNIVERSAL_APP_SHELL_LOADER}?type=web!${appEntry}`);
-
-  config.optimization
-    .minimizer('uglify')
-      .use(UglifyJSPlugin, [{
-        cache: true,
-        sourceMap: true,
-      }])
-      .end()
-    .minimizer('optimizeCSS')
-      .use(OptimizeCSSAssetsPlugin, [{
-        canPrint: true,
-      }]);
+  setUserConfig(config, context, 'web');
 };
